Separate files `RProps` and `RState`
