+++ kotlin-react/src/main/kotlin/react/Component.ext.kt
@@ -1,0 +1,7 @@
+package react
+
+import kotlinext.js.assign
+
+fun <S : RState> Component<*, S>.setState(buildState: S.() -> Unit) {
+    setState({ assign(it, buildState) })
+}
