+++ kotlin-react/src/main/kotlin/react/ErrorInfo.kt
@@ -1,0 +1,6 @@
+package react
+
+external interface RErrorInfo
+
+val RErrorInfo.componentStack: Any
+    get() = asDynamic().componentStack
