+++ kotlin-react/src/main/kotlin/react/Props.kt
@@ -1,8 +1,5 @@
 package react
 
-import kotlinext.js.assign
-
-// Props
 external interface RProps
 
 val RProps.children: Any
@@ -25,18 +22,3 @@ var RProps.ref: RRef
 fun <T> RProps.ref(ref: (T?) -> Unit) {
     asDynamic().ref = ref
 }
-
-// State
-external interface RState
-
-class BoxedState<T>(var state: T) : RState
-
-// Error info
-external interface RErrorInfo
-
-val RErrorInfo.componentStack: Any
-    get() = asDynamic().componentStack
-
-fun <S : RState> Component<*, S>.setState(buildState: S.() -> Unit) =
-    setState({ assign(it, buildState) })
-
