+++ kotlin-react/src/main/kotlin/react/State.kt
@@ -1,0 +1,5 @@
+package react
+
+external interface RState
+
+class BoxedState<T>(var state: T) : RState
